refactor: 밋팀 관리 페이지 설정 내의 라디오 박스 스타일드 컴포넌트 위치 변경
